test: extract Rocq generation support
