+++ tests/src/rocq_typecheck.rs
@@ -31,9 +31,9 @@ mod gate {
     use crate::rocq_decls::{
         Tok, declared_names, ident_at, is_punct, strip_rocq_comments, tokenize,
     };
-    use crate::utils::{build_ast, get_test_data_path};
-    use inference_type_checker::TypeCheckerBuilder;
-    use inference_wasm_codegen::{CompilationMode, OptLevel, Target};
+    use crate::rocq_test_support::{self, generate_v};
+    use crate::utils::get_test_data_path;
+    use inference_wasm_codegen::CompilationMode;
     use rustc_hash::{FxHashMap, FxHashSet};
     use std::path::{Path, PathBuf};
     use std::process::Command;
@@ -178,7 +178,11 @@ mod gate {
                 ExternalBytes::Fixture {
                     source,
                     module_name,
-                } => compile_fixture(source, module_name, CompilationMode::Compile),
+                } => rocq_test_support::compile_fixture(
+                    source,
+                    module_name,
+                    CompilationMode::Compile,
+                ),
                 ExternalBytes::Artifact { file } => {
                     let path = get_test_data_path().join("wasmlib").join(file);
                     std::fs::read(&path).unwrap_or_else(|e| panic!("read {}: {e}", path.display()))
@@ -459,35 +463,21 @@ mod gate {
         ("ge_u", "(ROI_ge SX_U)"),
     ];
 
-    /// Compiles one fixture under `tests/test_data/inf/` and returns its WASM.
-    fn compile_fixture(file: &str, module_name: &str, mode: CompilationMode) -> Vec<u8> {
-        let path = get_test_data_path().join("inf").join(file);
-        let source = std::fs::read_to_string(&path)
-            .unwrap_or_else(|e| panic!("read {}: {e}", path.display()));
-        let arena = build_ast(source);
-        let typed_context = TypeCheckerBuilder::build_typed_context(arena)
-            .unwrap_or_else(|e| panic!("type check failed for {file}: {e}"))
-            .typed_context();
-        inference_wasm_codegen::codegen(
-            &typed_context,
-            module_name,
-            inference_wasm_codegen::CodegenOptions {
-                target: Target::Wasm32,
-                mode,
-                opt_level: OptLevel::O3,
-                features: inference_wasm_codegen::EmitFeatures::default(),
-                layout: inference_wasm_codegen::MemoryLayout::default(),
-            },
-        )
-        .unwrap_or_else(|e| panic!("codegen failed for {file}: {e}"))
-        .wasm()
-        .to_vec()
-    }
+    #[test]
+    fn shared_generator_preserves_existing_output() {
+        let existing_wasm = rocq_test_support::compile_fixture(
+            "rocq_exists_spec.inf",
+            "rocq_exists_spec",
+            CompilationMode::Proof,
+        );
+        let existing = rocq_test_support::translate(
+            "rocq_exists_spec.inf",
+            "rocq_exists_spec",
+            &existing_wasm,
+        );
+        let shared = rocq_test_support::generate_v("rocq_exists_spec.inf", "rocq_exists_spec");
 
-    /// Proof-mode `.v` for one single-file fixture, driven entirely in-process.
-    fn generate_v(file: &str, module_name: &str) -> String {
-        let wasm = compile_fixture(file, module_name, CompilationMode::Proof);
-        translate(file, module_name, &wasm)
+        assert_eq!(shared, existing);
     }
 
     /// Proof-mode `.v` for a fixture that links external `.wasm` modules,
@@ -507,7 +497,7 @@ mod gate {
             .iter()
             .map(|(name, bytes)| (*name, bytes.as_slice()))
             .collect();
-        let main = compile_fixture(file, module_name, CompilationMode::Proof);
+        let main = rocq_test_support::compile_fixture(file, module_name, CompilationMode::Proof);
         // The unchecked write-set mode: each external here is supplied as bytes
         // by the fixture itself rather than resolved from the program's `use`
         // clauses, so no declaration-derived contract is in hand. The subject is
@@ -515,21 +505,7 @@ mod gate {
         // it — the merged body has no imports left to describe.
         let linked = inference::link(&main, &lib_refs, None)
             .unwrap_or_else(|e| panic!("link failed for {file}: {e}"));
-        translate(file, module_name, &linked)
-    }
-
-    /// The translation step both generators share.
-    fn translate(file: &str, module_name: &str, wasm: &[u8]) -> String {
-        // Empty explicit maps: the per-spec indices and the hassert obligations
-        // both ride along in the embedded `inference.spec_funcs` /
-        // `inference.hspecs` custom sections (see ROCQ_CONTRACT.md). For a
-        // linked module they are also the only correct source — the linker
-        // rewrote the embedded indices into the post-merge space, leaving
-        // codegen's own record stale.
-        let empty: FxHashMap<String, Vec<u32>> = FxHashMap::default();
-        let empty_hspecs = inference::HSpecMap::default();
-        inference::wasm_to_v(module_name, wasm, &empty, &empty_hspecs)
-            .unwrap_or_else(|e| panic!("wasm_to_v failed for {file}: {e}"))
+        rocq_test_support::translate(file, module_name, &linked)
     }
 
     /// One generated module this suite hands to `coqc`.
